scraper/requirements.txt
gunicorn==19.9.0
djangorestframework==3.9.1
Django==2.1.7
psycopg2==2.7.5
psycopg2-binary==2.7.5

